[feat] ai token limiter plugin
