+++ shenyu-plugin/pom.xml
@@ -50,7 +50,7 @@
         <module>shenyu-plugin-security</module>
         <module>shenyu-plugin-fault-tolerance</module>
         <module>shenyu-plugin-ai-proxy</module>
-        <module>shenyu-plugin-ai-statistic</module>
+        <module>shenyu-plugin-ai-token-limiter</module>
     </modules>
 
     <dependencies>
