--- shenyu-plugin/shenyu-plugin-ai-statistic/src/main/java/org/apache/shenyu/plugin/ai/statistic/redis/serializer/ByteArrayRedisSerializer.java
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shenyu.plugin.ai.statistic.redis.serializer;
-
-import org.springframework.data.redis.serializer.RedisSerializer;
-import org.springframework.data.redis.serializer.SerializationException;
-import org.springframework.lang.Nullable;
-
-/**
- * ByteArrayRedisSerializer.
- */
-public class ByteArrayRedisSerializer implements RedisSerializer<byte[]> {
-
-    @Override
-    @Nullable
-    public byte[] serialize(final @Nullable byte[] bytes) throws SerializationException {
-        return bytes;
-    }
-
-    @Override
-    @Nullable
-    public byte[] deserialize(final @Nullable byte[] bytes) throws SerializationException {
-        return bytes;
-    }
-}
