+++ shenyu-spring-boot-starter/shenyu-spring-boot-starter-plugin/pom.xml
@@ -75,6 +75,6 @@
         <module>shenyu-spring-boot-starter-plugin-logging-huawei-lts</module>
         <module>shenyu-spring-boot-starter-plugin-basic-auth</module>
         <module>shenyu-spring-boot-starter-plugin-ai-proxy</module>
-        <module>shenyu-spring-boot-starter-plugin-ai-statistic</module>
+        <module>shenyu-spring-boot-starter-plugin-ai-token-limiter</module>
     </modules>
 </project>
